Adding noise LFO waveform and slew setting

+++ src/hw_lfo.cpp
@@ -10,7 +10,7 @@ void HW_LFO::update() {
 
 //TODO: figure out approximate LFO rates
 const char* HW_LFO::getRateStr(char* buf, size_t len, uint8_t v) {
-    float freq = MIN_HZ * pow(RANGE, v / 127.0);
+    float freq = MIN_HZ * powf(RANGE, v / 127.0);
 
     char floatBuffer[10]; //buffer for float to string
     const uint8_t decimalPlaces = (freq < 1) ? 3 : (freq < 10) ? 2 : 1;
